Various updates and fixes

+++ removeovsrproconf.sh
@@ -1,2 +1,5 @@
-rm /etc/ld.so.conf.d/ovsrpro.conf
-ldconfig
+if [[ $1 -eq 0 ]]
+then
+  rm /etc/ld.so.conf.d/ovsrpro.conf
+  ldconfig
+fi
